06/09-lugonzal-FINALCHECK=1

+++ Libft/include/libft.h
@@ -22,7 +22,7 @@ typedef struct s_list
 	struct s_list	*next;
 }	t_list;
 
-void	*ft_memset(void *str, int c, size_t len);
+void	*ft_memset(void *str, long c, size_t len);
 void	ft_bzero(void *str, size_t n);
 void	*ft_memcpy(void *dst, const void *src, size_t n);
 void	*ft_memccpy(void *dst, const void *src, int c, size_t n);
